Remove sensei-quiz block and move the logic to block-hooks
