--- wp-content/themes/pub/wporg-learn-2024/src/sensei-quiz/block.json
@@ -1,4 +1,0 @@
-{
-	"name": "wporg-learn/quiz",
-	"viewScript": "file:./view.js"
-}
